Misc cleanup to reduce intermediate parses; ensure PreviewPanel's syntax tree gets updated on rule _name_ changes.
 * Allow clean of embedded parser files
 * Allow clean / reset of entire JFS location stores
 * RebuildSubscriptions should double check that mapped JFS files really exist, and recreate them if necessary, before trying to run Antlr generation, in case something has cleaned them
 * Generation result no longer inadvertently deletes mapped grammar files from the JFS
 * JFS classloader and storage logging on create/close/dispose
 * Checksum token and rule names in extractor, so that we can detect if the set of rules have not changed but their names have when deciding to update the preview syntax tree and such
 * Rewrite GeneralHighlighter to be based on AbstractHighlighter, and ensure that every single highlighter for a mime type doesn't enqueue its own separate parse job for every edit - this generated lots of pointless duplicate work
 * Ensure GeneralHighlighter and friends are associated with the document _and text component_ so we don't wind up with one highlighter trying to service two editors and two carets
 * Bump version to 2.0.63

+++ jfs/src/main/java/org/nemesis/jfs/JFSFileObjectImpl.java
@@ -208,6 +208,9 @@ class JFSFileObjectImpl implements JFSFileObject {
 
     @Override
     public boolean delete() {
+        if (deleted) {
+            return false;
+        }
         storage.discard();
         boolean result = storage.storage().delete(name, storage);
         if (result) {
